move render_markers_symbolizer to separate .cpp
