+++ Makefile
@@ -41,6 +41,7 @@ src/json/libmapnik-json.a:
 	$(PYTHON) scons/scons.py -j1 \
 		--config=cache --implicit-cache --max-drift=1 \
 		src/renderer_common/render_group_symbolizer.os \
+		src/renderer_common/render_markers_symbolizer.os \
 		src/renderer_common/render_thunk_extractor.os \
 		src/json/libmapnik-json.a \
 		src/wkt/libmapnik-wkt.a \
@@ -48,9 +49,6 @@ src/json/libmapnik-json.a:
 		src/expression_grammar.os \
 		src/transform_expression_grammar.os \
 		src/image_filter_grammar.os \
-		src/agg/process_markers_symbolizer.os \
-		src/grid/process_markers_symbolizer.os \
-		src/cairo/process_markers_symbolizer.os \
 
 mapnik: src/json/libmapnik-json.a
 	# then install the rest with -j$(JOBS)
