+++ src/agg/process_markers_symbolizer.cpp
@@ -38,8 +38,9 @@
 #include <mapnik/svg/svg_path_attributes.hpp>
 #include <mapnik/symbolizer.hpp>
 #include <mapnik/parse_path.hpp>
-#include <mapnik/renderer_common/process_markers_symbolizer.hpp>
 #include <mapnik/renderer_common/clipping_extent.hpp>
+#include <mapnik/renderer_common/render_markers_symbolizer.hpp>
+
 // agg
 #include "agg_basics.h"
 #include "agg_renderer_base.h"
