+++ src/build.py
@@ -254,6 +254,7 @@ source = Split(
     color_factory.cpp
     renderer_common.cpp
     renderer_common/render_group_symbolizer.cpp
+    renderer_common/render_markers_symbolizer.cpp
     renderer_common/render_pattern.cpp
     renderer_common/render_thunk_extractor.cpp
     math.cpp
