+++ src/renderer_common/render_markers_symbolizer.cpp
@@ -2,7 +2,7 @@
  *
  * This file is part of Mapnik (c++ mapping toolkit)
  *
- * Copyright (C) 2015 Artem Pavlenko
+ * Copyright (C) 2016 Artem Pavlenko
  *
  * This library is free software; you can redistribute it and/or
  * modify it under the terms of the GNU Lesser General Public
@@ -20,23 +20,24 @@
  *
  *****************************************************************************/
 
-#ifndef MAPNIK_RENDERER_COMMON_PROCESS_MARKERS_SYMBOLIZER_HPP
-#define MAPNIK_RENDERER_COMMON_PROCESS_MARKERS_SYMBOLIZER_HPP
-
 #include <mapnik/svg/svg_storage.hpp>
 #include <mapnik/svg/svg_path_adapter.hpp>
 #include <mapnik/vertex_converters.hpp>
 #include <mapnik/marker_cache.hpp>
 #include <mapnik/marker_helpers.hpp>
 #include <mapnik/geometry_type.hpp>
+#include <mapnik/renderer_common/render_markers_symbolizer.hpp>
+#include <mapnik/symbolizer.hpp>
 
 namespace mapnik {
 
-template <typename VD, typename RD, typename RendererType, typename ContextType>
+namespace detail {
+
+template <typename Detector, typename RendererType, typename ContextType>
 struct render_marker_symbolizer_visitor
 {
-    using vector_dispatch_type = VD;
-    using raster_dispatch_type = RD;
+    using vector_dispatch_type = vector_markers_dispatch<Detector>;
+    using raster_dispatch_type = raster_markers_dispatch<Detector>;
 
     using vertex_converter_type = vertex_converter<clip_line_tag,
                                                    clip_poly_tag,
@@ -52,7 +53,7 @@ struct render_marker_symbolizer_visitor
                                      proj_transform const& prj_trans,
                                      RendererType const& common,
                                      box2d<double> const& clip_box,
-                                     ContextType const& renderer_context)
+                                     ContextType & renderer_context)
         : filename_(filename),
           sym_(sym),
           feature_(feature),
@@ -242,38 +243,57 @@ struct render_marker_symbolizer_visitor
     proj_transform const& prj_trans_;
     RendererType const& common_;
     box2d<double> const& clip_box_;
-    ContextType const& renderer_context_;
+    ContextType & renderer_context_;
 };
 
-template <typename RendererType>
+} // namespace detail
+
+markers_dispatch_params::markers_dispatch_params(box2d<double> const& size,
+                                                 agg::trans_affine const& tr,
+                                                 symbolizer_base const& sym,
+                                                 feature_impl const& feature,
+                                                 attributes const& vars,
+                                                 double scale,
+                                                 bool snap)
+    : placement_params{
+        .size = size,
+        .tr = tr,
+        .spacing = get<value_double, keys::spacing>(sym, feature, vars),
+        .max_error = get<value_double, keys::max_error>(sym, feature, vars),
+        .allow_overlap = get<value_bool, keys::allow_overlap>(sym, feature, vars),
+        .avoid_edges = get<value_bool, keys::avoid_edges>(sym, feature, vars),
+        .direction = get<direction_enum, keys::direction>(sym, feature, vars)}
+    , placement_method(get<marker_placement_enum, keys::markers_placement_type>(sym, feature, vars))
+    , ignore_placement(get<value_bool, keys::ignore_placement>(sym, feature, vars))
+    , snap_to_pixels(snap)
+    , scale_factor(scale)
+    , opacity(get<value_double, keys::opacity>(sym, feature, vars))
+{
+    placement_params.spacing *= scale;
+}
+
 void render_markers_symbolizer(markers_symbolizer const& sym,
                                mapnik::feature_impl & feature,
                                proj_transform const& prj_trans,
-                               RendererType const& common,
+                               renderer_common const& common,
                                box2d<double> const& clip_box,
                                markers_renderer_context & renderer_context)
 {
     using Detector = decltype(*common.detector_);
-    using VD = vector_markers_dispatch<Detector>;
-    using RD = raster_markers_dispatch<Detector>;
-    using ContextType = markers_renderer_context & ;
+    using RendererType = renderer_common;
+    using ContextType = markers_renderer_context;
+    using VisitorType = detail::render_marker_symbolizer_visitor<Detector,
+                                                                 RendererType,
+                                                                 ContextType>;
 
-    using namespace mapnik::svg;
     std::string filename = get<std::string>(sym, keys::file, feature, common.vars_, "shape://ellipse");
     if (!filename.empty())
     {
-        std::shared_ptr<mapnik::marker const> mark = mapnik::marker_cache::instance().find(filename, true);
-        render_marker_symbolizer_visitor<VD,RD,RendererType,ContextType> visitor(filename,
-                                                                                 sym,
-                                                                                 feature,
-                                                                                 prj_trans,
-                                                                                 common,
-                                                                                 clip_box,
-                                                                                 renderer_context);
+        auto mark = mapnik::marker_cache::instance().find(filename, true);
+        VisitorType visitor(filename, sym, feature, prj_trans, common, clip_box,
+                            renderer_context);
         util::apply_visitor(visitor, *mark);
     }
 }
 
 } // namespace mapnik
-
-#endif // MAPNIK_RENDERER_COMMON_PROCESS_MARKERS_SYMBOLIZER_HPP
