+++ src/renderer_common/render_thunk_extractor.cpp
@@ -23,8 +23,7 @@
 // mapnik
 #include <mapnik/label_collision_detector.hpp>
 #include <mapnik/make_unique.hpp>
-#include <mapnik/marker_helpers.hpp>
-#include <mapnik/renderer_common/process_markers_symbolizer.hpp>
+#include <mapnik/renderer_common/render_markers_symbolizer.hpp>
 #include <mapnik/renderer_common/render_thunk_extractor.hpp>
 
 namespace mapnik {
